requirements.txt
# Core runtime packages
python-dateutil>=2.8.0
requests>=2.25.0

# HTML processing (for LinkedIn company page parsing)
beautifulsoup4==4.12.3
lxml==5.3.0

# Date/timezone handling
pytz==2024.2

# Environment variables for local development
python-dotenv>=0.19.0

# System monitoring for Lambda optimization
psutil==5.9.8

# Regex and text processing (built-in modules: re, html)
# urllib.parse (built-in)

# Concurrency (built-in concurrent.futures)
# typing (built-in for Python 3.9+)
# logging (built-in)
# time (built-in)
# os (built-in)
# json (built-in)
